requirements.txt
nltk==3.8.1
openai-whisper==20230314
pandas==2.0.0
pydub==0.25.1
scikit-learn==1.2.2
moviepy==1.0.3